requirements.txt
python-dotenv==1.1.1
semantic-kernel==1.34.0
openai==1.93.1